Extract identity-owners composition hook from useAppShellComposition
